possible ubuntu fixes for uint64 and SDL2
Former-commit-id: f0062d256d76e04ed42d9d82d7e403d3e4a13b6b

+++ src/my_sys.h
@@ -139,7 +139,7 @@ typedef unsigned int         uint32_t;
 
 #ifndef _UINT64_T
 #define _UINT64_T
-typedef unsigned long long   uint64_t;
+typedef unsigned long int   uint64_t;
 #endif // _UINT64_T
 #endif
